refactor: rename service context file
